Move global 'copyClipboard' to 'installation.js'
