+++ src/js/entry.js
@@ -1,5 +1,4 @@
-const {buildMenuTwisties, copyClipboard, persistUrlQuery} = require('./common');
-global.copyClipboard = copyClipboard;
+const {buildMenuTwisties, persistUrlQuery} = require('./common');
 
 const {selectLatestPlatform, unselectLatestPlatform} = require('./releases');
 Object.assign(global, {selectLatestPlatform, unselectLatestPlatform});
